# Reduce state machines in binary expressions