--- a/Fluid/Ast/BinaryExpression.cs
+++ b/Fluid/Ast/BinaryExpression.cs
@@ -1,4 +1,9 @@
-﻿namespace Fluid.Ast
+﻿using System;
+using System.Runtime.CompilerServices;
+using System.Threading.Tasks;
+using Fluid.Values;
+
+namespace Fluid.Ast
 {
     public abstract class BinaryExpression : Expression
     {
@@ -11,5 +16,38 @@
         public Expression Left { get; }
 
         public Expression Right { get; }
+
+        /// <summary>
+        /// Evaluates two operands and tries to avoid state machines.
+        /// </summary>
+        public override ValueTask<FluidValue> EvaluateAsync(TemplateContext context)
+        {
+            var leftTask = Left.EvaluateAsync(context);
+            var rightTask = Right.EvaluateAsync(context);
+
+            if (leftTask.IsCompletedSuccessfully && rightTask.IsCompletedSuccessfully)
+            {
+                return Evaluate(leftTask.Result, rightTask.Result);
+            }
+
+            return Awaited(leftTask, rightTask);
+        }
+
+        [MethodImpl(MethodImplOptions.NoInlining)]
+        private async ValueTask<FluidValue> Awaited(
+            ValueTask<FluidValue> leftTask,
+            ValueTask<FluidValue> rightTask)
+        {
+            var leftValue = await leftTask;
+            var rightValue = await rightTask;
+
+            return Evaluate(leftValue, rightValue);
+        }
+
+        // sub-classes using the default implementation need to override this
+        internal virtual FluidValue Evaluate(FluidValue leftValue, FluidValue rightValue)
+        {
+            throw new NotImplementedException();
+        }
     }
 }