--- a/Fluid/Ast/BinaryExpressions/AndBinaryExpression.cs
+++ b/Fluid/Ast/BinaryExpressions/AndBinaryExpression.cs
@@ -1,33 +1,16 @@
-﻿using System.Threading.Tasks;
-using Fluid.Values;
+﻿using Fluid.Values;
 
 namespace Fluid.Ast.BinaryExpressions
 {
-    public class AndBinaryExpression : BinaryExpression
+    public sealed class AndBinaryExpression : BinaryExpression
     {
         public AndBinaryExpression(Expression left, Expression right) : base(left, right)
         {
         }
 
-        public override ValueTask<FluidValue> EvaluateAsync(TemplateContext context)
+        internal override FluidValue Evaluate(FluidValue leftValue, FluidValue rightValue)
         {
-            static async ValueTask<FluidValue> Awaited(ValueTask<FluidValue> leftTask, ValueTask<FluidValue> rightTask)
-            {
-                var leftValue = await leftTask;
-                var rightValue = await rightTask;
-
-                return BooleanValue.Create(leftValue.ToBooleanValue() && rightValue.ToBooleanValue());
-            }
-
-            var leftTask = Left.EvaluateAsync(context);
-            var rightTask = Right.EvaluateAsync(context);
-
-            if (leftTask.IsCompletedSuccessfully && rightTask.IsCompletedSuccessfully)
-            {
-                return BooleanValue.Create(leftTask.Result.ToBooleanValue() && rightTask.Result.ToBooleanValue());
-            }
-
-            return Awaited(leftTask, rightTask);
+            return BooleanValue.Create(leftValue.ToBooleanValue() && rightValue.ToBooleanValue());
         }
     }
 }