--- a/Fluid/Ast/BinaryExpressions/DivideBinaryExpression.cs
+++ b/Fluid/Ast/BinaryExpressions/DivideBinaryExpression.cs
@@ -1,19 +1,15 @@
-﻿using System.Threading.Tasks;
-using Fluid.Values;
+﻿using Fluid.Values;
 
 namespace Fluid.Ast.BinaryExpressions
 {
-    public class DivideBinaryExpression : BinaryExpression
+    public sealed class DivideBinaryExpression : BinaryExpression
     {
         public DivideBinaryExpression(Expression left, Expression right) : base(left, right)
         {
         }
 
-        public override async ValueTask<FluidValue> EvaluateAsync(TemplateContext context)
+        internal override FluidValue Evaluate(FluidValue leftValue, FluidValue rightValue)
         {
-            var leftValue = await Left.EvaluateAsync(context);
-            var rightValue = await Right.EvaluateAsync(context);
-
             if (leftValue is NumberValue && rightValue is NumberValue)
             {
                 var rightNumber = rightValue.ToNumberValue();