--- a/Fluid/Ast/BinaryExpressions/ModuloBinaryExpression.cs
+++ b/Fluid/Ast/BinaryExpressions/ModuloBinaryExpression.cs
@@ -1,25 +1,18 @@
-﻿using System.Threading.Tasks;
-using Fluid.Values;
+﻿using Fluid.Values;
 
 namespace Fluid.Ast.BinaryExpressions
 {
-    public class ModuloBinaryExpression : BinaryExpression
+    public sealed class ModuloBinaryExpression : BinaryExpression
     {
         public ModuloBinaryExpression(Expression left, Expression right) : base(left, right)
         {
         }
 
-        public override async ValueTask<FluidValue> EvaluateAsync(TemplateContext context)
+        internal override FluidValue Evaluate(FluidValue leftValue, FluidValue rightValue)
         {
-            var leftValue = await Left.EvaluateAsync(context);
-            var rightValue = await Right.EvaluateAsync(context);
-
-            if (leftValue is NumberValue && rightValue is NumberValue)
-            {
-                return NumberValue.Create(leftValue.ToNumberValue() % rightValue.ToNumberValue());
-            }
-
-            return NilValue.Instance;
+            return leftValue is NumberValue && rightValue is NumberValue
+                ? NumberValue.Create(leftValue.ToNumberValue() % rightValue.ToNumberValue())
+                : NilValue.Instance;
         }
     }
 }