--- a/Fluid/Ast/BinaryExpressions/MultiplyBinaryExpression.cs
+++ b/Fluid/Ast/BinaryExpressions/MultiplyBinaryExpression.cs
@@ -1,25 +1,18 @@
-﻿using System.Threading.Tasks;
-using Fluid.Values;
+﻿using Fluid.Values;
 
 namespace Fluid.Ast.BinaryExpressions
 {
-    public class MultiplyBinaryExpression : BinaryExpression
+    public sealed class MultiplyBinaryExpression : BinaryExpression
     {
         public MultiplyBinaryExpression(Expression left, Expression right) : base(left, right)
         {
         }
 
-        public override async ValueTask<FluidValue> EvaluateAsync(TemplateContext context)
+        internal override FluidValue Evaluate(FluidValue leftValue, FluidValue rightValue)
         {
-            var leftValue = await Left.EvaluateAsync(context);
-            var rightValue = await Right.EvaluateAsync(context);
-
-            if (leftValue is NumberValue && rightValue is NumberValue)
-            {
-                return NumberValue.Create(leftValue.ToNumberValue() * rightValue.ToNumberValue());
-            }
-
-            return NilValue.Instance;
+            return leftValue is NumberValue && rightValue is NumberValue
+                ? NumberValue.Create(leftValue.ToNumberValue() * rightValue.ToNumberValue())
+                : NilValue.Instance;
         }
     }
 }