--- a/Fluid/Ast/BinaryExpressions/NotEqualBinaryExpression.cs
+++ b/Fluid/Ast/BinaryExpressions/NotEqualBinaryExpression.cs
@@ -1,25 +1,18 @@
-﻿using System.Threading.Tasks;
-using Fluid.Values;
+﻿using Fluid.Values;
 
 namespace Fluid.Ast.BinaryExpressions
 {
-    public class NotEqualBinaryExpression : BinaryExpression
+    public sealed class NotEqualBinaryExpression : BinaryExpression
     {
         public NotEqualBinaryExpression(Expression left, Expression right) : base(left, right)
         {
         }
 
-        public override async ValueTask<FluidValue> EvaluateAsync(TemplateContext context)
+        internal override FluidValue Evaluate(FluidValue leftValue, FluidValue rightValue)
         {
-            var leftValue = await Left.EvaluateAsync(context);
-            var rightValue = await Right.EvaluateAsync(context);
-
-            if (leftValue.Equals(rightValue))
-            {
-                return BooleanValue.False;
-            }
-
-            return BooleanValue.True;
+            return leftValue.Equals(rightValue)
+                ? BooleanValue.False 
+                : BooleanValue.True;
         }
     }
 }