--- a/Fluid/Ast/BinaryExpressions/SubstractBinaryExpression.cs
+++ b/Fluid/Ast/BinaryExpressions/SubstractBinaryExpression.cs
@@ -1,5 +1,4 @@
-﻿using System.Threading.Tasks;
-using Fluid.Values;
+﻿using Fluid.Values;
 
 namespace Fluid.Ast.BinaryExpressions
 {
@@ -9,17 +8,11 @@ namespace Fluid.Ast.BinaryExpressions
         {
         }
 
-        public override async ValueTask<FluidValue> EvaluateAsync(TemplateContext context)
+        internal override FluidValue Evaluate(FluidValue leftValue, FluidValue rightValue)
         {
-            var leftValue = await Left.EvaluateAsync(context);
-            var rightValue = await Right.EvaluateAsync(context);
-
-            if (leftValue is NumberValue && rightValue is NumberValue)
-            {
-                return NumberValue.Create(leftValue.ToNumberValue() - rightValue.ToNumberValue());
-            }
-
-            return NilValue.Instance;
+            return leftValue is NumberValue && rightValue is NumberValue
+                ? NumberValue.Create(leftValue.ToNumberValue() - rightValue.ToNumberValue())
+                : NilValue.Instance;
         }
     }
 }