Spotted some things when profiling Orchard Agency, so let's work the magic.

* Use generic pattern in BinaryExpression to reduce state machines
* improve FluidValue conversion
* remove state machine in FluidParserExtensions

## Fluid.Benchmarks.TagBenchmarks

| **Diff**|Method|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |BinaryExpressions_Parse|10.022 μs|0.0235 μs|0.8392|0.0000|0.0000|3560 B|
| **New** |	| **9,969.8 ns (-3%)** | **32.87 ns** | **0.8392 (0%)** | **0.0000** | **0.0000** | **3560 B (0%)** |
| Old |BinaryExpressions_Render|1.352 μs|0.0032 μs|0.0134|0.0000|0.0000|56 B|
| **New** |	| **971.0 ns (-30%)** | **3.56 ns** | **0.0134 (0%)** | **0.0000** | **0.0000** | **56 B (0%)** |